gcs_to_bq/requirements.txt
functions-framework==3.*
google-cloud-storage==2.17.0
google-cloud-bigquery==3.24.0